pyproject.toml
[tool.poetry]
name = "zoomcamp-project"
version = "0.1.0"
description = ""
authors = ["mf-andres <amunoz@gradiant.org>"]
readme = "README.md"
packages = [{include = "zoomcamp_project"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.5.3"
typer = "^0.7.0"
pyarrow = "^11.0.0"
selenium = "^4.8.3"
google-cloud-storage = "^2.7.0"
google-auth = "^2.16.3"
google-cloud-bigquery = "^3.9.0"
cryptography = "40.0.1"
dbt-bigquery = "^1.4.3"
pydantic = {extras = ["dotenv"], version = "^1.10.7"}
prefect = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
